Cleanup exports and README for v1.0.2 release
