+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "@alch/alchemy-sdk",
   "version": "1.0.1",
-  "description": "",
-  "author": "",
+  "description": "Extended Ethers.js SDK for Alchemy APIs",
+  "author": "Alchemy",
   "license": "MIT",
   "scripts": {
     "pre-commit": "lint-staged",
@@ -42,7 +42,7 @@
       "yarn format-file"
     ],
     "package.json": [
-      "./inject-version.js",
+      "node ./inject-version.js",
       "git add src/version.ts"
     ]
   },
